+++ src/main/java/data/concretesources/sql/SqlLocationDatasource.java
@@ -1,4 +1,4 @@
-package data.concretesources;
+package data.concretesources.sql;
 
 import com.google.inject.Inject;
 import data.abstractsources.LocationDatasource;
